[refactor] 상태 변화 시 다시 그리는 함수들 renderView에 합침
